Changes to constraints in Plan and addition of Nuclease Class table

+++ impc_prod_tracker/src/main/java/uk/ac/ebi/impc_prod_tracker/data/biology/guide/Guide.java
@@ -41,7 +41,7 @@ public class Guide extends BaseEntity
 
     private Boolean truncatedGuide;
 
-    private Integer grnaConcentration;
+    private Double grnaConcentration;
 
     private String pam3;
